adding multiple city functionality fixed bug
