+++ src/reducers/reducers-weather.js
@@ -4,12 +4,7 @@ const weatherReducer = function(state = [], action) {
   switch (action.type) {
     case FETCH_WEATHER:
       const cityName = action.payload.data.city;
-      const cityData = function (dataType) {
-      action.payload.data.list.map(function (c) {
-        return c.main[dataType]
-        })
-      }
-      return  {
+      const addCity = {
         city: cityName.name,
         id: cityName.id,
         lat: cityName.coord.lat,
@@ -24,6 +19,7 @@ const weatherReducer = function(state = [], action) {
           return c.main.humidity
         }),
       }
+      return  [addCity, ...state]
     default:
       return state;
   }
